Commit to move current models and controllers to data-structure folder
maintain(controllers/*, models/*): move controllers and models to data-structure
- move all models and controller files to data-structure folder
- modify path to the moved files in all dependent modules
- [Partly Delivers #165437339]
