+++ data-structure/models/user.js
@@ -1,6 +1,6 @@
 const { List, fromJS } = require('immutable');
-const userData = require('../data/user');
-const { asyncHashPassword } = require('../utilities/password');
+const userData = require('../../data/user');
+const { asyncHashPassword } = require('../../utilities/password');
 
 const userSchema = {
   id: 'number',
